+++ spec/spec_helper.rb
@@ -30,6 +30,7 @@ RSpec.configure do |config|
     base_url = Ferrum::Server.server.base_url
     options = { base_url: base_url }
     options.merge!(headless: false) if ENV["HEADLESS"] == "false"
+    options.merge!(slowmo: ENV["SLOWMO"].to_f) if ENV["SLOWMO"].to_f > 0
 
     if ENV["CI"]
       ferrum_logger = StringIO.new
@@ -53,7 +54,7 @@ RSpec.configure do |config|
   end
 
   config.after(:each) do |example|
-    save_exception_artifacts(browser, example.metadata, ferrum_logger) if ENV["CI"] && example.exception
+    save_exception_artifacts(browser, example.metadata, ferrum_logger) if ENV.fetch("CI", nil) && example.exception
 
     reset
   end